Setup basic game logic
https://www.gaya-game.com/collections/strategy-game/products/black-sheep

Nodes in the game board were named alphabetically starting left to right
and then top to bottom:

A --- B --- C
| \ / | \ / |
|  D  |  E  |
| / \ | / \ |
F --- G --- H
| \ / | \ / |
|  I  |  J  |
| / \ | / \ |
K --- L --- M

The term neighbor refers to a jumped node. toNode and destination are
used to refer to the jumped to node. So a move jumpping from A to G
would have a neighborNode of D, a destinationNode of G and a fromNode of
A.