Show win and lose states
Why:
When a player has either lost or won the game, we would like to show a
win or lose message to the player.

This commit:
Adds the logic for determining if the game is won by seeing if there is
only one dot left.
Adds the logic for determing if the game is lost by looking to see if
there is more than one dot left and no dot has a valid move. Currently
the noValidMoves for a dot logic is a bit cumbersome and could likely be
improved if we adjusted the model to lend itself better to folds.