Add levels screen
Why:
We would like for users to see a list of all the levels on a new screen.

This commit:
Pulls out the game logic to a game screen file and introduces the basic
logic for switching screens. Given this is intended to be a mobile app,
it didn't make too much sense to me to add all of the logic for
handleing url based web routing that is common in elm SPAs.

I'd much rather have a more mobile style of stack/tab/swtich navigation,
since that is what is expected on a mobile experience. The solution here
is far from that but I wanted to avoid as much complexity as possible at
this point. Given that this is a PWA and ultimately targets safari iOS
and chrome, this may be entirely the wrong direction.

Currently going to let Main.elm handle the current screen and navigation
elements and try to keep screen specific logic in screen specific files.
Again, this may be the wrong direction to go down.